requirements.txt
Flask
requests
PyMuPDF
reportlab
Pillow
streamlit
